BREAKING CHANGE: "servers" parameter changed to single "server" and added a new broadcast wrapper class for multiple servers
